Move loading of Mysql2Proxy to its own railtie
