+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Add ability to require only core files
 - Move loading of PostgreSQLProxy to its own Railtie
+- Move loading of Mysql2Proxy to its own Railtie
 
 ## [0.8.0] - 2025-08-28
