+++ lib/active_record_proxy_adapters/connection_handling.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
 require "active_record_proxy_adapters/connection_handling/postgresql_proxy"
-require "active_record_proxy_adapters/connection_handling/mysql2"
+require "active_record_proxy_adapters/connection_handling/mysql2_proxy"
 require "active_record_proxy_adapters/connection_handling/trilogy"
 require "active_record_proxy_adapters/connection_handling/sqlite3"
