+++ lib/active_record_proxy_adapters/connection_handling/mysql2_proxy.rb
